#405, #419: Добавлена проверка окружения, исправление ошибкок

+++ protected/modules/yupe/components/YWebModule.php
@@ -41,6 +41,10 @@ abstract class YWebModule extends CWebModule
      *  @var array опции редактора
      */
     public $editorOptions  = array();
+    /**
+     *  @var bool разрещение на вывод YFlashMessages
+     */
+    public $flashMess = false;
 
     /**
      *  @return string текущая версия модуля
@@ -210,6 +214,12 @@ abstract class YWebModule extends CWebModule
         return true;
     }
 
+    protected function setFlash($type, $message)
+    {
+        if ($this->flashMess)
+            Yii::app()->user->setFlash($type, $message);
+    }
+
     /**
      *  @return bool включает модуль
      *  @since 0.5
@@ -220,18 +230,26 @@ abstract class YWebModule extends CWebModule
         $fileModule = $yupe->getModulesConfigDefault($this->id);
         $fileConfig = $yupe->getModulesConfig($this->id);
 
-        if (@copy($fileModule, $fileConfig))
-            Yii::app()->user->setFlash(
+        if (is_file($fileConfig))
+            $this->setFlash(
                 YFlashMessages::NOTICE_MESSAGE,
-                Yii::t('yupe', 'Модуль включен!')
+                Yii::t('yupe', 'Модуль уже включен!')
             );
         else
         {
-            Yii::app()->user->setFlash(
-                YFlashMessages::ERROR_MESSAGE,
-                Yii::t('yupe', 'Произошла ошибка при включении модуля, конфигурационный файл поврежден или отсутствует доступ к папке config!')
-            );
-            return false;
+            if (@copy($fileModule, $fileConfig))
+                $this->setFlash(
+                    YFlashMessages::NOTICE_MESSAGE,
+                    Yii::t('yupe', 'Модуль включен!')
+                );
+            else
+            {
+                $this->setFlash(
+                    YFlashMessages::ERROR_MESSAGE,
+                    Yii::t('yupe', 'Произошла ошибка при включении модуля, конфигурационный файл поврежден или отсутствует доступ к папке config!')
+                );
+                return false;
+            }
         }
         return true;
     }
@@ -246,33 +264,41 @@ abstract class YWebModule extends CWebModule
         $fileModule = $yupe->getModulesConfigDefault($this->id);
         $fileConfig = $yupe->getModulesConfig($this->id);
 
+        if (!is_file($fileConfig))
+        {
+            $this->setFlash(
+                YFlashMessages::NOTICE_MESSAGE,
+                Yii::t('yupe', 'Модуль уже отключен!')
+            );
+            return true;
+        }
         if (!$this->isNoDisable)
         {
             if (@md5_file($fileModule) != @md5_file($fileConfig))
             {
                 $fileConfigBack = $yupe->getModulesConfigBack($this->id);
                 if (!@copy($fileConfig, $fileConfigBack))
-                    Yii::app()->user->setFlash(
+                    $this->setFlash(
                         YFlashMessages::ERROR_MESSAGE,
                         Yii::t('yupe', 'Произошла ошибка при копировании старого конфигурационного файла в папку modulesBack!')
                     );
             }
             if (@unlink($fileConfig))
             {
-                Yii::app()->user->setFlash(
+                $this->setFlash(
                     YFlashMessages::NOTICE_MESSAGE,
                     Yii::t('yupe', 'Модуль отключен!')
                 );
                 return true;
             }
             else
-                Yii::app()->user->setFlash(
+                $this->setFlash(
                     YFlashMessages::ERROR_MESSAGE,
                     Yii::t('yupe', 'Произошла ошибка при отключении модуля, нет доступа к конфигурационному файлу!')
                 );
         }
         else
-            Yii::app()->user->setFlash(
+            $this->setFlash(
                 YFlashMessages::ERROR_MESSAGE,
                 Yii::t('yupe', 'Этот модуль запрещено отключать!')
             );
